Feature be0039 config validation exception and response data

+++ src/main/java/com/smart/tailor/utils/response/MaterialResponse.java
@@ -18,7 +18,7 @@ public class MaterialResponse {
 
     private String categoryName;
 
-    private Double hsCode;
+    private Long hsCode;
 
     private String unit;
